# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Talabati water delivery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /driver/status:
    get:
      operationId: getDriverStatus
      tags: [driver]
      summary: Get all driver statuses
      responses:
        "200":
          description: Driver statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DriverStatus"
    post:
      operationId: updateDriverStatus
      tags: [driver]
      summary: Update driver availability status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverStatusInput"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverStatus"
  /driver/{driverId}/subscription:
    get:
      operationId: getDriverSubscription
      tags: [driver]
      summary: Get latest subscription payment status for driver
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Latest subscription payment or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionPayment"
        "404":
          description: No payment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: submitSubscriptionReceipt
      tags: [driver]
      summary: Submit a subscription payment receipt
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitReceiptInput"
      responses:
        "201":
          description: Receipt submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionPayment"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /driver/{driverId}/account:
    get:
      operationId: getDriverAccount
      tags: [driver]
      summary: Get driver account status and subscription info
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Driver account info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverAccount"
        "404":
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateDriverAccount
      tags: [driver]
      summary: Update driver account status or subscription (admin)
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverAccountInput"
      responses:
        "200":
          description: Driver account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverAccount"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /driver/{driverId}/docs:
    post:
      operationId: submitDriverDocs
      tags: [driver]
      summary: Save driver verification document URLs (truck video, photos, license)
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverDocsInput"
      responses:
        "200":
          description: Documents saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverDocsResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /driver/{driverId}/orders:
    get:
      operationId: getDriverOrders
      tags: [driver]
      summary: Get orders assigned to a specific driver
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Driver's assigned orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Order:
      type: object
      required: [id, userId, waterVolume, barrelCount, totalPrice, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        driverId:
          type: ["string", "null"]
        userName:
          type: ["string", "null"]
        userPhone:
          type: ["string", "null"]
        waterVolume:
          type: string
        barrelCount:
          type: integer
        totalPrice:
          type: number
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        status:
          type: string
        createdAt:
          type: string
...
    DriverStatus:
      type: object
      required: [driverId, driverName, currentStatus, updatedAt]
      properties:
        driverId:
          type: string
        driverName:
          type: string
        currentStatus:
          type: string
        updatedAt:
          type: string
    DriverStatusInput:
      type: object
      required: [driverId, currentStatus]
      properties:
        driverId:
          type: string
        currentStatus:
          type: string
          enum: ["حاضر", "استراحة", "مغلق"]
    DriverAccount:
      type: object
      required: [accountStatus, subscriptionExpired]
      properties:
        accountStatus:
          type: string
          enum: ["pending", "active"]
        subscriptionExpiresAt:
          type: ["string", "null"]
        subscriptionExpired:
          type: boolean
    UpdateDriverAccountInput:
      type: object
      properties:
        accountStatus:
          type: string
          enum: ["pending", "active"]
        subscriptionExpiresAt:
          type: ["string", "null"]
    SubscriptionPayment:
      type: object
      required: [id, driverId, status, createdAt]
      properties:
        id:
          type: string
        driverId:
          type: string
        receiptImage:
          type: string
        status:
          type: string
          enum: ["pending", "approved", "rejected"]
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
        reviewedAt:
          type: ["string", "null"]
    SubmitReceiptInput:
      type: object
      required: [receiptImage]
      properties:
        receiptImage:
          type: string
    DriverDocsInput:
      type: object
      required: [truckFrontPhotoUrl, driverLicenseUrl]
      properties:
        truckFrontPhotoUrl:
          type: string
        driverLicenseUrl:
          type: string
        truckVideoUrl:
          type: string
          description: "اختياري — لم يعد مطلوباً (تمت إزالة فيديو التحقق)"
        truckSidePhotoUrl:
          type: string
          description: "اختياري — لم يعد مطلوباً (تمت إزالة صورة الجانب)"
    DriverDocsResponse:
      type: object
      required: [driverId]
      properties:
        driverId:
          type: string
        truckVideoUrl:
          type: ["string", "null"]
        truckFrontPhotoUrl:
          type: ["string", "null"]
        truckSidePhotoUrl:
          type: ["string", "null"]
        driverLicenseUrl:
          type: ["string", "null"]